Error handling delete command

+++ cogs/commands/info/userinfo.py
@@ -187,9 +187,10 @@ class UserInfo(commands.Cog):
 
         """
 
-        if not self.bot.settings.permissions.hasAtLeast(ctx.guild, ctx.author, 5):
+        bot_chan = self.bot.settings.guild().channel_botspam
+        if not self.bot.settings.permissions.hasAtLeast(ctx.guild, ctx.author, 5) and ctx.channel.id != bot_chan:
             await ctx.message.delete()
-            raise commands.BadArgument("You do not have permission to use this command.")
+            raise commands.BadArgument(f"Command only allowed in <#{bot_chan}>")
 
         results = await self.bot.settings.user(user.id)
         
@@ -218,9 +219,9 @@ class UserInfo(commands.Cog):
         """
 
         await ctx.message.delete()
-        if not self.bot.settings.permissions.hasAtLeast(ctx.guild, ctx.author, 5):
-            await ctx.message.delete()
-            raise commands.BadArgument("You do not have permission to use this command.")
+        bot_chan = self.bot.settings.guild().channel_botspam
+        if not self.bot.settings.permissions.hasAtLeast(ctx.guild, ctx.author, 5) and ctx.channel.id != bot_chan:
+            raise commands.BadArgument(f"Command only allowed in <#{bot_chan}>")
         
         if isinstance(user, int):
             user = await self.bot.fetch_user(user)
